new branch: feat: updating configuration set up
splitting out config files to be separated based on use case (shared, user deployment) for security and admin overhead reduction. Intention will be to allow cross platform configurations to be isolated to admins/root/service accounts with keys stored more securely such that they're not sourced into every shell... sorry about that lol. Was green in the ears when initially set that up.